+++ app/src/main/java/com/unison/appartment/activities/SignInActivity.java
@@ -1,29 +1,24 @@
 package com.unison.appartment.activities;
 
-import androidx.annotation.NonNull;
-
 import android.os.Bundle;
 import android.view.View;
 import android.widget.EditText;
 
-import com.google.android.gms.tasks.OnCompleteListener;
-import com.google.android.gms.tasks.Task;
 import com.google.android.material.floatingactionbutton.FloatingActionButton;
 import com.google.android.material.textfield.TextInputLayout;
-import com.google.firebase.auth.AuthResult;
 import com.google.firebase.auth.FirebaseAuthInvalidCredentialsException;
 import com.google.firebase.auth.FirebaseAuthInvalidUserException;
-import com.google.firebase.database.DataSnapshot;
 import com.google.firebase.database.DatabaseError;
-import com.google.firebase.database.DatabaseReference;
-import com.google.firebase.database.ValueEventListener;
 import com.unison.appartment.Appartment;
 import com.unison.appartment.database.Auth;
 import com.unison.appartment.database.AuthListener;
-import com.unison.appartment.database.Database;
-import com.unison.appartment.database.DatabaseListener;
+import com.unison.appartment.database.DatabaseReader;
+import com.unison.appartment.database.DatabaseReaderListener;
+import com.unison.appartment.database.DatabaseWriter;
+import com.unison.appartment.database.DatabaseWriterListener;
 import com.unison.appartment.database.FirebaseAuth;
-import com.unison.appartment.database.FirebaseDatabase;
+import com.unison.appartment.database.FirebaseDatabaseReader;
+import com.unison.appartment.database.FirebaseDatabaseWriter;
 import com.unison.appartment.fragments.FirebaseProgressDialogFragment;
 import com.unison.appartment.model.User;
 import com.unison.appartment.utils.KeyboardUtils;
@@ -35,7 +30,7 @@ import com.unison.appartment.R;
 public class SignInActivity extends FormActivity {
 
     private Auth auth;
-    private Database database;
+    private DatabaseReader databaseReader;
 
     EditText inputEmail;
     EditText inputPassword;
@@ -48,7 +43,7 @@ public class SignInActivity extends FormActivity {
         setContentView(R.layout.activity_sign_in);
 
         auth = new FirebaseAuth();
-        database = new FirebaseDatabase();
+        databaseReader = new FirebaseDatabaseReader();
 
         // Precondizione: Se viene acceduta questa activity, vuol dire che non c'è nessun utente loggato
 
@@ -83,10 +78,14 @@ public class SignInActivity extends FormActivity {
             public void onClick(View v) {
                 KeyboardUtils.hideKeyboard(SignInActivity.this);
                 if (checkInput()) {
-                    auth.performSignIn(inputEmail.getText().toString(), inputPassword.getText().toString(), new AuthListener() {
+                    progressDialog = FirebaseProgressDialogFragment.newInstance(
+                            getString(R.string.activity_signin_progress_title),
+                            getString(R.string.activity_signin_progress_description));
+                    progressDialog.show(getSupportFragmentManager(), FirebaseProgressDialogFragment.TAG_FIREBASE_PROGRESS_DIALOG);
+                    auth.signIn(inputEmail.getText().toString(), inputPassword.getText().toString(), new AuthListener() {
                         @Override
                         public void onSuccess() {
-                            database.retrieveUser(auth.getCurrentUserUid(), new DatabaseListener() {
+                            databaseReader.retrieveUser(auth.getCurrentUserUid(), new DatabaseReaderListener() {
                                 @Override
                                 public void onReadSuccess(Object object) {
                                     Appartment.getInstance().setUser((User) object);
@@ -106,16 +105,6 @@ public class SignInActivity extends FormActivity {
                                     la situazione non può essere risolta dall'utente.
                                     */
                                     showErrorDialog();
-                                }
-
-                                @Override
-                                public void onWriteSuccess() {
-
-                                }
-
-                                @Override
-                                public void onWriteFail(Exception exception) {
-
                                 }
                             });
                             moveToNextActivity(UserProfileActivity.class);
@@ -176,33 +165,5 @@ public class SignInActivity extends FormActivity {
 
         return result;
     }
-
-    /*private void retrieveUser(String uid) {
-        String separator = getString(R.string.db_separator);
-        String path = getString(R.string.db_users) + separator + getString(R.string.db_users_uid, uid);
-        DatabaseReference dbRef = FirebaseDatabase.getInstance().getReference(path);
-        dbRef.addListenerForSingleValueEvent(new ValueEventListener() {
-            @Override
-            public void onDataChange(@NonNull DataSnapshot dataSnapshot) {
-                if (dataSnapshot.exists()) {
-                    Appartment.getInstance().setUser(dataSnapshot.getValue(User.class));
-                }
-                else {
-                    // TODO Gestire l'errore
-                }
-            }
-
-            @Override
-            public void onCancelled(@NonNull DatabaseError databaseError) {
-                *//*
-                onCancelled viene invocato solo se si verifica un errore a lato server oppure se
-                le regole di sicurezza impostate in Firebase non permettono l'operazione richiesta.
-                In questo caso perciò viene visualizzato un messaggio di errore generico, dato che
-                la situazione non può essere risolta dall'utente.
-                 *//*
-                showErrorDialog();
-            }
-        });
-    }*/
 
 }
