+++ app/src/main/java/com/unison/appartment/database/DatabaseReader.java
@@ -1,0 +1,5 @@
+package com.unison.appartment.database;
+
+public interface DatabaseReader {
+    void retrieveUser(final String uid, final DatabaseReaderListener listener);
+}
